% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_pt_data.R
\name{proc_pt_data}
\alias{proc_pt_data}
\title{Process proficiency test data from multiple sources}
\usage{
proc_pt_data(df)
}
\arguments{
\item{df}{data frame with PT data}
}
\value{
dataframe
}
\description{
Process proficiency test data from multiple sources
}
