% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{run_app}
\usage{
run_app(app = "dashboard_test")
}
\arguments{
\item{app}{name of app to run. Options: \code{dashboard_test}, \code{upload_test}}
}
\value{
shiny app executes
}
\description{
run_app
}
